refactor: share mixed attachment E2E setup
